requirements.txt
pandas~=2.2.1
qrcode~=7.4.2
pillow~=10.3.0
openpyxl~=3.1.2